+++ cnccode/data_classes/DBEPortalCustomerDocumentWithoutFile.inc.php
@@ -3,6 +3,7 @@
 * @authors Karim Ahmed
 * @access public
 */
+global $cfg;
 require_once($cfg["path_gc"] . "/DBEntity.inc.php");
 
 class DBEPortalCustomerDocumentWithoutFile extends DBEntity
@@ -10,8 +11,7 @@ class DBEPortalCustomerDocumentWithoutFile extends DBEntity
     const portalCustomerDocumentID = "portalCustomerDocumentID";
     const customerID = "customerID";
     const description = "description";
-    const startersFormFlag = "startersFormFlag";
-    const leaversFormFlag = "leaversFormFlag";
+    const customerContract = "customerContract";
     const mainContactOnlyFlag = "mainContactOnlyFlag";
     const createdDate = "createdDate";
     const createdUserID = "createdUserID";
@@ -21,6 +21,7 @@ class DBEPortalCustomerDocumentWithoutFile extends DBEntity
      * @access public
      * @param void
      * @return void
+     * @throws Exception
      * @see constructor()
      */
     function __construct(&$owner)
@@ -30,8 +31,7 @@ class DBEPortalCustomerDocumentWithoutFile extends DBEntity
         $this->addColumn(self::portalCustomerDocumentID, DA_ID, DA_NOT_NULL);
         $this->addColumn(self::customerID, DA_ID, DA_ALLOW_NULL);
         $this->addColumn(self::description, DA_STRING, DA_NOT_NULL);
-        $this->addColumn(self::startersFormFlag, DA_YN, DA_NOT_NULL);
-        $this->addColumn(self::leaversFormFlag, DA_YN, DA_NOT_NULL);
+        $this->addColumn(self::customerContract, DA_BOOLEAN, DA_NOT_NULL, null, 0);
         $this->addColumn(self::mainContactOnlyFlag, DA_YN, DA_NOT_NULL);
         $this->addColumn(
             self::createdDate,
@@ -45,7 +45,14 @@ class DBEPortalCustomerDocumentWithoutFile extends DBEntity
         $this->setAddColumnsOff();
     }
 
-
+    public function hasContractDocumentByCustomerId($customerID)
+    {
+        $queryString = 'select count(*) > 0 from ' . $this->getTableName() . " where " . $this->getDBColumnName(
+                self::customerContract
+            ) . " = 1 and " . $this->getDBColumnName(self::customerID) . " = $customerID";
+        $result = $this->db->query($queryString);
+        return +$result->fetch_row()[0];
+    }
 }
 
 ?>