+++ htdocs/Office365LicensesExport.php
@@ -286,14 +286,6 @@ do {
                 "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
             );
             $dbeCustomerDocument->setValue(
-                DBEPortalCustomerDocument::startersFormFlag,
-                'N'
-            );
-            $dbeCustomerDocument->setValue(
-                DBEPortalCustomerDocument::leaversFormFlag,
-                'N'
-            );
-            $dbeCustomerDocument->setValue(
                 DBEPortalCustomerDocument::mainContactOnlyFlag,
                 'Y'
             );